requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
analytics-python==1.2.9   # via -r requirements/base.txt
appdirs==1.4.4            # via virtualenv
argparse==1.4.0           # via -r requirements/base.txt, stevedore
astroid==2.3.3            # via pylint, pylint-celery
attrs==20.2.0             # via -r requirements/base.txt, edx-ace, pytest
bok-choy==1.1.1           # via -r requirements/test.in
certifi==2020.6.20        # via -r requirements/base.txt, requests
cffi==1.14.3              # via -r requirements/base.txt, cryptography
chardet==3.0.4            # via -r requirements/base.txt, requests
click-log==0.3.2          # via edx-lint
click==7.1.2              # via click-log, edx-lint
coreapi==2.3.3            # via -r requirements/base.txt, django-rest-swagger, openapi-codec
coreschema==0.0.4         # via -r requirements/base.txt, coreapi
coverage==5.3             # via -r requirements/test.in
cryptography==3.1.1       # via -r requirements/base.txt, pyjwt
ddt==1.4.1                # via -r requirements/test.in
defusedxml==0.6.0         # via -r requirements/base.txt, python3-openid, social-auth-core
distlib==0.3.1            # via virtualenv
django-appconf==1.0.4     # via -r requirements/base.txt, django-statici18n
django-compat==1.0.15     # via -r requirements/base.txt, django-hijack
django-debug-toolbar==3.1.1  # via -r requirements/base.txt
django-extensions==3.0.9  # via -r requirements/base.txt
django-filter==2.4.0      # via -r requirements/base.txt
django-hijack==2.1.10     # via -r requirements/base.txt
django-ratelimit==3.0.1   # via -r requirements/base.txt
django-rest-swagger==2.2.0  # via -r requirements/base.txt
django-sortedm2m==3.0.2   # via -r requirements/base.txt
django-statici18n==2.0.0  # via -r requirements/base.txt
django-storages==1.8      # via -c requirements/constraints.txt, -r requirements/base.txt
django-waffle==2.0.0      # via -r requirements/base.txt, edx-django-utils, edx-drf-extensions
django-webpack-loader==0.7.0  # via -r requirements/base.txt
djangorestframework==3.12.1  # via -r requirements/base.txt, django-rest-swagger, drf-jwt, edx-drf-extensions, rest-condition
drf-jwt==1.17.2           # via -r requirements/base.txt, edx-drf-extensions
edx-ace==0.1.15           # via -r requirements/base.txt
edx-auth-backends==3.1.0  # via -r requirements/base.txt
edx-django-release-util==0.4.4  # via -r requirements/base.txt
edx-django-sites-extensions==2.5.1  # via -r requirements/base.txt
edx-django-utils==3.8.0   # via -r requirements/base.txt, edx-drf-extensions, edx-rest-api-client
edx-drf-extensions==6.2.0  # via -r requirements/base.txt
edx-lint==1.5.2           # via -r requirements/test.in
edx-opaque-keys==2.1.1    # via -r requirements/base.txt, edx-drf-extensions
edx-rest-api-client==5.2.1  # via -r requirements/base.txt
git+https://github.com/edx/credentials-themes.git@0.1.30#egg=edx_credentials_themes==0.1.30  # via -r requirements/base.txt
factory-boy==3.1.0        # via -r requirements/test.in
faker==4.5.0              # via factory-boy
filelock==3.0.12          # via tox, virtualenv
future==0.18.2            # via -r requirements/base.txt, pyjwkest
httpretty==1.0.2          # via -r requirements/test.in
idna==2.10                # via -r requirements/base.txt, requests
iniconfig==1.0.1          # via pytest
isort==4.2.5              # via -c requirements/constraints.txt, -r requirements/test.in, pylint
itypes==1.2.0             # via -r requirements/base.txt, coreapi
jinja2==2.11.2            # via -r requirements/base.txt, coreschema
lazy-object-proxy==1.4.3  # via astroid
lazy==1.4                 # via bok-choy
markdown==3.3             # via -r requirements/base.txt
markupsafe==1.1.1         # via -r requirements/base.txt, jinja2
mccabe==0.6.1             # via pylint
mysqlclient==2.0.1        # via -r requirements/base.txt
newrelic==5.20.1.150      # via -r requirements/base.txt, edx-django-utils
oauthlib==3.1.0           # via -r requirements/base.txt, requests-oauthlib, social-auth-core
openapi-codec==1.3.2      # via -r requirements/base.txt, django-rest-swagger
packaging==20.4           # via pytest, tox
pbr==5.5.0                # via -r requirements/base.txt, stevedore
pep8==1.7.1               # via -r requirements/test.in
pillow==7.2.0             # via -r requirements/base.txt
pluggy==0.13.1            # via pytest, tox
psutil==5.7.2             # via -r requirements/base.txt, edx-django-utils
py==1.9.0                 # via pytest, tox
pycparser==2.20           # via -r requirements/base.txt, cffi
pycryptodomex==3.9.8      # via -r requirements/base.txt, pyjwkest
pygments==2.7.1           # via -r requirements/base.txt
pyjwkest==1.4.2           # via -r requirements/base.txt, edx-drf-extensions
pyjwt[crypto]==1.7.1      # via -r requirements/base.txt, drf-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pylint-celery==0.3        # via edx-lint
pylint-django==2.0.11     # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.4.4             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pymongo==3.11.0           # via -r requirements/base.txt, edx-opaque-keys
pyparsing==2.4.7          # via packaging
pytest-django==3.10.0     # via -r requirements/test.in
pytest==6.1.1             # via -r requirements/test.in, pytest-django
python-dateutil==2.8.1    # via -r requirements/base.txt, analytics-python, edx-ace, edx-drf-extensions, faker
python-memcached==1.59    # via -r requirements/base.txt
python3-openid==3.2.0     # via -r requirements/base.txt, social-auth-core
pytz==2020.1              # via -r requirements/base.txt, django
pyyaml==5.3.1             # via -r requirements/base.txt, edx-django-release-util
requests-oauthlib==1.3.0  # via -r requirements/base.txt, social-auth-core
requests==2.24.0          # via -r requirements/base.txt, analytics-python, coreapi, edx-drf-extensions, edx-rest-api-client, pyjwkest, requests-oauthlib, responses, sailthru-client, slumber, social-auth-core
responses==0.12.0         # via -r requirements/test.in
rest-condition==1.0.3     # via -r requirements/base.txt, edx-drf-extensions
sailthru-client==2.2.3    # via -r requirements/base.txt, edx-ace
selenium==3.141.0         # via bok-choy
semantic-version==2.8.5   # via -r requirements/base.txt, edx-drf-extensions
simplejson==3.17.2        # via -r requirements/base.txt, django-rest-swagger, sailthru-client
six==1.15.0               # via -r requirements/base.txt, analytics-python, astroid, bok-choy, cryptography, django-compat, django-statici18n, edx-ace, edx-auth-backends, edx-django-release-util, edx-drf-extensions, edx-lint, edx-opaque-keys, packaging, pyjwkest, python-dateutil, python-memcached, responses, social-auth-app-django, social-auth-core, stevedore, tox, virtualenv
slumber==0.7.1            # via -r requirements/base.txt, edx-rest-api-client
git+https://github.com/python-social-auth/social-app-django.git@c00d23c2b45c3317bd35b15ad1b959338689cef8#egg=social-auth-app-django  # via -r requirements/base.txt, edx-auth-backends
social-auth-core==3.2.0   # via -c requirements/constraints.txt, -r requirements/base.txt, edx-auth-backends, social-auth-app-django
sqlparse==0.4.1           # via -r requirements/base.txt, django, django-debug-toolbar
stevedore==1.10.0         # via -c requirements/constraints.txt, -r requirements/base.txt, edx-ace, edx-django-utils, edx-opaque-keys
testfixtures==6.14.2      # via -r requirements/test.in
text-unidecode==1.3       # via faker
toml==0.10.1              # via pytest, tox
tox==3.20.0               # via -r requirements/test.in
uritemplate==3.0.1        # via -r requirements/base.txt, coreapi
urllib3==1.25.10          # via -r requirements/base.txt, requests, responses, selenium
virtualenv==20.0.33       # via tox
wrapt==1.11.2             # via astroid
xss-utils==0.1.3          # via -r requirements/base.txt
